refactor: split useSetting internals
- move persistence and updater helpers into dedicated module files

- relocate update tests alongside implementation

Generated-by: Codex (model: gpt-5-codex)
